Moved showImage from HomeFunctions to GameFunctions because it is a core function and was not making sense have it in that separated way
