--- app/src/main/java/ar/teamrocket/duelosmeli/domain/HomeFunctions.kt
@@ -1,9 +1,0 @@
-package ar.teamrocket.duelosmeli.domain
-
-import android.content.Context
-import android.widget.ImageView
-
-interface HomeFunctions {
-
-    fun showImage(context: Context, image: Int, imageView: ImageView)
-}
